Abstract TypeTableHeader and add grouping
